Default ScriptScope storage is now LuaTable
Added an extension method which allows you to create a "LuaScope", which
is a ScriptScope object using a LuaTable as the backing store for
variables. This allows the use of metatables, and metamethods and
generally allows the ScriptScope to behave as the Lua _G object should.

Updated tests and other methods to support this, as well as adding
custom support for this particular scope type into LuaContext for
[Get,TryGet,Get<T> and Set]ScopeVariable operations.