Continuing work on CodeContext Implementation
Currently working on getting IronLua back to a compilable (and runnable)
state. It currently passes most tests, but fails on syntax trees where
NaN checking is done (since the LuaRuntimeException constructors are
marked as internal. We will need to fix that before we can continue.
When that is done, then we need to add support for rewriting lambda
expressions for debug operations, and run some tests for closure support
(should work given how the DLR is written, out of the box, but we will
need to see). Also need to add tests for custom environments for
different functions (it is now present in CodeContext, and just needs to
be enabled in BaseLibrary).

Then I need to clean up the code, and remove some expression types which
I added but am not using. Maybe also add some new expression types to
represent different things which may change, and which we can optimize
the Expression tree by formatting depending on those changes.