Continuing work on CodeContext Implementation
Fully functional now, all tests passing as they should as well as some
new tests to ensure that closures and function environments are working
correctly. We still need to get debug tracing working correctly, and
implement the DebugContext handlers before we have most of the TODO
stuff completed, but the backend is now correctly rewritten.