Performance Optimizations
Added some compile-time optimizations for global member access (in
particular, determining at compile time whether an identifier represents
a native library or not). This has shaved a massive amount of time off
of our global member accesses, particularly on the first run (which is
now almost as fast a subsequent runs).